structex

Module re

The required interface for an underlying regex engine

Structs

Captures

Represents the capture group positions for a single RegexEngine match in terms of byte offsets into the original haystack that the match was run against.

RawCaptures

Represents the capture group positions for a single RegexEngine match only in terms of byte offsets into the original haystack.

Traits

Haystack

A haystack that can be searched by a RegexEngine.

RegexEngine

A RegexEngine is an underlying regular expression engine that can be used to match and extract text as part of a structural regular expression.

Sliceable

Something that supports extracting a contiguous sub-section between two bytes offsets.

Writable

Something that can be written to to a given io::Write.