structex::re

Trait Haystack

pub trait Haystack:
    Writable
    + Debug
    + PartialEq
    + Eq
    + Copy {
    type Slice<'h>: Writable
       where Self: 'h;

    // Required methods
    fn slice<'h>(&'h self, from: usize, to: usize) -> Self::Slice<'h>;
    fn max_len(&self) -> usize;
}

A haystack is an associated type on Re that the regular expression engine can be run against.

Typically this is a &str but some engines may support richer types in order to provide searching of streams or discontiguous inputs.

Required Associated Types

type Slice<'h>: Writable where Self: 'h

The output of the slice method.

Typically the same type as the haystack itself but not required to be so.

Required Methods

fn slice<'h>(&'h self, from: usize, to: usize) -> Self::Slice<'h>

A contiguous sub-section of the haystack between the given bytes offsets.

The given byte offsets from a half-open interval, inclusive of from but omitting to. This is the same semantics as a normal Rust range from..to.

fn max_len(&self) -> usize

The maximum length of the full haystack in bytes.

This value will be used as the upper bound to extract slices from this haystack when searching for matches. As such, this value must be a valid to argument to the slice method.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.