structex::re

Trait Re

pub trait Re: Sized {
    type CompileError: Error + 'static;
    type Haystack: Haystack<Slice = Self::Slice> + ?Sized;
    type Slice: Writable + ?Sized;

    // Required methods
    fn compile(re: &str) -> Result<Self, Self::CompileError>;
    fn is_match_between(
        &self,
        haystack: &Self::Haystack,
        from: usize,
        to: usize,
    ) -> bool;
    fn captures_between(
        &self,
        haystack: &Self::Haystack,
        from: usize,
        to: usize,
    ) -> Option<RawCaptures>;
}

An Re is an underlying regular expression engine that can be used to match and extract text as part of a structural regular expression.

The interface provided by this trait is intended for use by Structex as an internal implementation detail and is likely not particularly useful outside of this crate. Implementors of this trait should pay particular attention to the requirements around compilation in order for the resulting regex expressions to be compatible with this crate.

Required Associated Types

type CompileError: Error + 'static

Error type that is returned by the Re::compile method when compilation of a given regular expression fails. This will be wrapped into an Error when returned from Structex::compile or StructexBuilder::build.

type Haystack: Haystack<Slice = Self::Slice> + ?Sized

The haystack that can be searched by this Re using the is_match_between and captures_between methods.

type Slice: Writable + ?Sized

The slice type of the associated Haystack that is returned by Captures methods when extracting matches and submatches.

Required Methods

fn compile(re: &str) -> Result<Self, Self::CompileError>

Attempt to compile the given regular expression for use inside of a Structex.

Multi-line matching

Almost all use cases for structural regular expressions involve matching patterns that span multiple lines. As such, any underlying regular expression engine used with Structex must support multi-line patterns in order to provide the expected semantics of each of the different operators that build upon it.

See here for information on how this is implemented in the regex crate for reference.

fn is_match_between( &self, haystack: &Self::Haystack, from: usize, to: usize, ) -> bool

Returns true if there is a match for the regex between the given byte offsets in the haystack.

This does not need to search for the leftmost-longest match and where possible should be faster to run that Re::captures_between which needs to extract the position of the match itself and all submatches.

fn captures_between( &self, haystack: &Self::Haystack, from: usize, to: usize, ) -> Option<RawCaptures>

Searches for the first match of this regex between the given byte offsets in the given haystack, returning the overall match along with the matches of each capture group in the regex. If no match is found, then None is returned.

See RawCaptures::new for requirements around constructing the return type.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.