structex::re

Trait Sliceable

pub trait Sliceable: Writable {
    type Slice<'h>: Writable
       where Self: 'h;

    // Required methods
    fn char_at(&self, byte_offset: usize) -> Option<char>;
    fn slice<'h>(&'h self, range: Range<usize>) -> Self::Slice<'h>;
    fn max_len(&self) -> usize;
}

Something that supports extracting a contiguous sub-section between two bytes offsets.

Required Associated Types

type Slice<'h>: Writable where Self: 'h

The output of the slice method.

Required Methods

fn char_at(&self, byte_offset: usize) -> Option<char>

Returns the character at the given byte offset if one exists, otherwise None.

fn slice<'h>(&'h self, range: Range<usize>) -> Self::Slice<'h>

The contiguous sub-section of self that is denoted by the given byte Range.

fn max_len(&self) -> usize

The maximum length in bytes.

This value will be used as the upper bound to extract slices when searching for matches. As such, this value must be a valid end to the range argument to the slice method.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.