structex

Module template

Simple template parsing and rendering for use in actions.

This module provides a simple Template type that can be parsed from the argument string of a given Action and used for rendering string content based on submatches extracted by a Structex.

Syntax

The syntax supported for templating is extremely minimal and focuses on injecting submatches into a user provided template. In its simplest form a template is simply a string literal, but the syntax also supports referencing submatches by their index. To inject the contents of a submatch at a particular point within a template, place the capture index inside of curly braces: "submatch 1 is {1} and submatch 2 is {2}".

It is also permitted to reference arbitrary variables within a template provided that template is rendered with either the render_with_context or render_with_context_to methods.

Syntax Errors

It is an error to have an unclosed {} pair within a template. To escape a curly brace, place a \ before the opening brace. Closing braces do not need to be escaped.

use structex::template::Template;

// The following are all valid templates

// String literals with no references are allowed
assert!(Template::parse("hello, world!").is_ok());

// Raw strings with `\n` and `\t` escape sequences
assert!(Template::parse(r#"hello,\tworld!\n"#).is_ok());

// Escaping `{` requires `\\` to provide a literal backslash
assert!(Template::parse("\\{ 1, 2, 3 }").is_ok());

// Using a raw string is easier
assert!(Template::parse(r#"\{ "hello", "world!" }"#).is_ok());

// A single reference
assert!(Template::parse("hello, {1}!").is_ok());

// Multiple references
assert!(Template::parse("{2}, {1}!").is_ok());

// The same reference multiple times
assert!(Template::parse("{1}, {1}!").is_ok());

// Something other than a capture index as a reference
assert!(Template::parse("{foo}").is_ok());

// The following are all invalid templates

// An unclosed capture reference
assert!(Template::parse("{1").is_err());

// An unknown escape sequence
assert!(Template::parse("\\[").is_err());

Usage

To make use of Templates with the rest of this crate, construct your templates from the parsed Actions returned by a Structex instance and then perform your rendering based on the matches that are yielded from Structex::iter_tagged_captures.

use structex::{Structex, template::Template};
use regex::Regex;

let haystack = r#"This is a multi-line
string that mentions peoples names.
People like Alice and Bob. People
like Claire and David, but really
we're here to talk about Alice.
Alice is everyone's friend."#;

let se: Structex<Regex> = Structex::new(r#"
  x/(.|\n)*?\./ {
    g/Alice/ n/(\w+)\./ p/The last word is '{1}'/;
    v/Alice/ n/(\w+)/   p/The first word is '{1}'/;
  }
"#).unwrap();

// Parse and register the templates
let templates: Vec<Template> = se
    .actions()
    .iter()
    .map(|action| Template::parse(action.arg().unwrap()).unwrap())
    .collect();

let output: Vec<String> = se
    .iter_tagged_captures(haystack)
    .map(|caps| {
        let id = caps.id().unwrap();
        templates[id].render(&caps).unwrap()
    })
    .collect();

assert_eq!(
    &output,
    &[
        "The first word is 'This'",
        "The last word is 'Bob'",
        "The last word is 'Alice'",
        "The last word is 'friend'",
    ]
);

To render values other than captures from a Structex, implement the Context trait:

use structex::{Structex, template::{Context, RenderError, Template}};
use regex::Regex;
use std::io::{self, Write};

struct Ctx;

impl Context for Ctx {
    fn render_var<W>(&self, var: &str, w: &mut W) -> Option<io::Result<usize>>
    where
        W: io::Write
    {
        match var {
            "foo" => Some(w.write(b"last word")),
            "bar" => Some(w.write(b"first word")),
            _ => None
        }
    }
}

let haystack = r#"This is a multi-line
string that mentions peoples names.
People like Alice and Bob. People
like Claire and David, but really
we're here to talk about Alice.
Alice is everyone's friend."#;

let se: Structex<Regex> = Structex::new(r#"
  x/(.|\n)*?\./ {
    g/Alice/ n/(\w+)\./ p/The {foo} is '{1}'/;
    v/Alice/ n/(\w+)/   p/The {bar} is '{1}'/;
  }
"#).unwrap();

// Parse and register the templates
let templates: Vec<Template> = se
    .actions()
    .iter()
    .map(|action| Template::parse(action.arg().unwrap()).unwrap())
    .collect();

let output: Vec<String> = se
    .iter_tagged_captures(haystack)
    .map(|caps| {
        let id = caps.id().unwrap();
        templates[id].render_with_context(&caps, &Ctx).unwrap()
    })
    .collect();

assert_eq!(
    &output,
    &[
        "The first word is 'This'",
        "The last word is 'Bob'",
        "The last word is 'Alice'",
        "The last word is 'friend'",
    ]
);

Structs

Template

A parsed string template that can be rendered for a given TaggedCaptures.

Enums

ErrorKind

A list specifying the different categories of errors that can be encountered while parsing a template.

RenderError

An error that can occur whilst rendering a Template.

Traits

Context

A Context is a type that can be passed to a Template’s render_with_context or render_with_context_to methods in order to resolve and render variable references beyond TaggedCaptures submatch indices.

Type Aliases

Error

An error that can arise during template parsing.