structex::re

Trait Haystack

pub trait Haystack<R>: Sliceable
where
    R: RegexEngine,
{
    // Required methods
    fn is_match_between(&self, re: &R, from: usize, to: usize) -> bool;
    fn captures_between(
        &self,
        re: &R,
        from: usize,
        to: usize,
    ) -> Option<RawCaptures>;
}

A haystack that can be searched by a RegexEngine.

Typically this is a &str but some engines may support richer types in order to provide searching of streams or discontiguous inputs.

Required Methods

fn is_match_between(&self, re: &R, from: usize, to: usize) -> bool

Returns true if there is a match for the regex between the given byte offsets in the haystack.

This does not need to search for the leftmost-longest match and where possible should be faster to run that Haystack::captures_between which needs to extract the position of the match itself and all submatches.

fn captures_between( &self, re: &R, from: usize, to: usize, ) -> Option<RawCaptures>

Searches for the first match of this regex between the given byte offsets in the given haystack, returning the overall match along with the matches of each capture group in the regex. If no match is found, then None is returned.

See RawCaptures::new for requirements around constructing the return type.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.